Split and refactor
